html/canvas/offscreen/text/2d.text.style.spacing.and.canvas.spacing.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.text.style.spacing.and.canvas.spacing</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.style.spacing.and.canvas.spacing</h1>
<p class="desc">Testing letter spacing and word spacing work with canvas Style</p>


<script>
var t = async_test("Testing letter spacing and word spacing work with canvas Style");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

var canvas = new OffscreenCanvas(100, 50);
var ctx = canvas.getContext('2d');

_assertSame(ctx.letterSpacing, '0px', "ctx.letterSpacing", "'0px'");
_assertSame(ctx.wordSpacing, '0px', "ctx.wordSpacing", "'0px'");
var width_normal = ctx.measureText('letter spacing').width;

// There are 14 letters in "letter spacing", sothe length is increased by 28.
canvas.style.letterSpacing = '2px';
var new_width = ctx.measureText('letter spacing').width;
_assertSame(new_width, width_normal + 28, "new_width", "width_normal + 28");

// there are 2 words in "letter spacing", so the length is increased by 4
canvas.style.wordSpacing = '2px';
canvas.style.letterSpacing = '0px';
new_width = ctx.measureText('letter spacing').width;
_assertSame(new_width, width_normal + 4, "new_width", "width_normal + 4");
t.done();

});
</script>
